+++ core-bundle/src/DependencyInjection/Configuration.php
@@ -101,6 +101,8 @@ class Configuration implements ConfigurationInterface
                         ->end()
                     ->end()
                 ->end()
+                ->variableNode('localconfig')
+                ->end()
             ->end()
         ;