+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 4.3.0-beta1 (2016-XX-XX)
 
+ * Allow to adjust the Contao configuration by adding settings under "contao.localconfig" (see #521).
  * Make script combination optional in case the website supports HTTP/2 (see #484).
  * Automatically find the template files.
  * Make the legacy configuration files optional (see #521).